Refactored redis code to make it easier for other views to use the functionality.
Fixed a bug that caused the re-fetch behavior to work backwards. Instead of re-fetching results that were originaly fetched more than a day ago, the results were being fetched again if the results had been fetched less than a day ago; potentially leading to a never expiring cache and most likely to a hit to the feed libraries on every request.

Added tests to ensure the redis cache behavior operates as expected.